+++ countUp.js
@@ -2,7 +2,7 @@
 
     countUp.js
     by @inorganik
-    v 0.0.5
+    v 0.0.6
 
 */
 
@@ -41,23 +41,21 @@ function countUp(target, startVal, endVal, decimals, duration) {
     }
 
     var self = this;
-    
-    this.d = document.getElementById(target);
-    
-    startVal = Number(startVal);
-    endVal = Number(endVal);
-    this.countDown = (startVal > endVal) ? true : false;
-    
+
     // toggle easing
     this.useEasing = true;
     
+    this.d = document.getElementById(target);
+    self.startVal = Number(startVal);
+    endVal = Number(endVal);
+    this.countDown = (startVal > endVal) ? true : false;
     decimals = Math.max(0, decimals || 0);
     this.dec = Math.pow(10, decimals);
     this.duration = duration * 1000 || 2000;
-
     this.startTime = null;
+    this.timestamp = null;
+    this.remaining = null;
     this.frameVal = startVal;
-
     this.rAF = null;
     
     // Robert Penner's easeOutExpo
@@ -67,23 +65,26 @@ function countUp(target, startVal, endVal, decimals, duration) {
     this.count = function(timestamp) {
         
         if (self.startTime === null) self.startTime = timestamp;
-     
+
+        self.timestamp = timestamp;
+
         var progress = timestamp - self.startTime;
+        self.remaining = self.duration - progress;
         
         // to ease or not to ease
         if (self.useEasing) {
             if (self.countDown) {
-                var i = self.easeOutExpo(progress, 0, startVal - endVal, self.duration);
+                var i = self.easeOutExpo(progress, 0, self.startVal - endVal, self.duration);
                 self.frameVal = startVal - i;
             } else {
-                self.frameVal = self.easeOutExpo(progress, startVal, endVal - startVal, self.duration);
+                self.frameVal = self.easeOutExpo(progress, self.startVal, endVal - self.startVal, self.duration);
             }
         } else {
             if (self.countDown) {
-                var i = (startVal - endVal) * (progress / self.duration);
-                self.frameVal = startVal - i;
+                var i = (self.startVal - endVal) * (progress / self.duration);
+                self.frameVal = self.startVal - i;
             } else {
-                self.frameVal = startVal + (endVal - startVal) * (progress / self.duration);
+                self.frameVal = self.startVal + (endVal - self.startVal) * (progress / self.duration);
             }
         }
         
@@ -111,7 +112,7 @@ function countUp(target, startVal, endVal, decimals, duration) {
         self.callback = callback;
         // make sure values are valid
         if (!isNaN(endVal) && !isNaN(startVal)) {
-            requestAnimationFrame(self.count);
+            self.rAF = requestAnimationFrame(self.count);
         } else {
             console.log('countUp error: startVal or endVal is not a number');
             self.d.innerHTML = '--';
@@ -123,7 +124,13 @@ function countUp(target, startVal, endVal, decimals, duration) {
     }
     this.reset = function() {
         cancelAnimationFrame(self.rAF);
-        this.d.innerHTML = startVal;
+        self.d.innerHTML = self.addCommas(startVal.toFixed(decimals));
+    }
+    this.resume = function() {
+        self.startTime = null;
+        self.duration = self.remaining;
+        self.startVal = self.frameVal;
+        requestAnimationFrame(self.count);
     }
     this.addCommas = function(nStr) {
         nStr += '';
@@ -137,6 +144,9 @@ function countUp(target, startVal, endVal, decimals, duration) {
         }
         return x1 + x2;
     }
+
+    // format startVal on initialization
+    self.d.innerHTML = self.addCommas(startVal.toFixed(decimals));
 }
 // Example:
 // var numAnim = new countUp("SomeElementYouWantToAnimate", 0, 99.99, 2, 1.5);