+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+### v 0.0.6 (1/17/14)
+
+- added resume method
+- format startVal on initialization
+
 ### v 0.0.5 (1/17/14)
 
 - added stop method